percentage complete is gucci

+++ src/main/java/com/dolphinblue/service/LessonService.java
@@ -29,9 +29,9 @@ public class LessonService {
         //Get tasks of lesson.
         List<Key<Task>> tasks = lesson.getTasks();
         //number of lessons.
-        int total = tasks.size();
+        double total = tasks.size();
         //number of lessons completed.
-        int count = 0;
+        double count = 0;
 
         for(int i = 0; i < tasks.size(); i++) {
             Key task_key = tasks.get(i);
